% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_age_mixing.R
\name{best_age_mixing}
\alias{best_age_mixing}
\title{Return best-fitting mixing structure}
\usage{
best_age_mixing(mixing_data, start_ages, max_age = 74)
}
\description{
A convenience function - calculates the likelihood of observing the provided data under each mixing structure provided in \link[mixage]{estimate_age_mixing}
}
