% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_single_year_age_distribution.R
\name{calculate_single_year_age_distribution}
\alias{calculate_single_year_age_distribution}
\title{Calculate age distribution.}
\usage{
calculate_single_year_age_distribution(min_age, max_model_age)
}
\arguments{
\item{min_age}{minimum model age}

\item{max_model_age}{maximum model age}
}
\value{
\code{age_prop}: the proportion of the model population with each age
}
\description{
Calculate an age distribution from annual probabilities of death (from life tables).
}
