% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_group_matrix.R
\name{define_age_group_matrix}
\alias{define_age_group_matrix}
\title{Create age mixing matrices for the desired age groups.}
\usage{
define_age_group_matrix(start_ages, max_age = 74)
}
\arguments{
\item{start_ages}{A numeric vector, where the ith entry is the youngest age in the ith age group}

\item{max_age}{The oldest age in the model population. The oldest age group will include the ages from \code{max(start_age)} to \code{max_age - 1}}
}
\description{
Uses the best estimates from A Journal Article, 2018, to define male and female age mixing matrices for the supplied age groups.
}
\details{
TODO: allow user to provide age distribution, calculate probabilities using LOTP
}
