% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_age_mixing.R
\name{estimate_age_mixing}
\alias{estimate_age_mixing}
\title{Calculate age mixing structures}
\usage{
estimate_age_mixing(choice_data, start_ages, distribution = c("gamma",
  "normal"), link = c("identity", "log"), max_age = 74,
  age_distribution = NULL)
}
\arguments{
\item{choice_data}{a dataframe with chooser age, partner age, sex, and optional survey weights. 
These columns should be named \code{chsage}, \code{ptage}, \code{sex}, and \code{weights}, respectively.
Sex must be coded as "Male" and "Female"}

\item{start_ages}{vector of the youngest ages included in each age group. 
If \code{start_ages} is \code{c(12, 20, 30)}, the age groups are 12-19, 20-29, and 30 to \code{max_age - 1}.}

\item{distribution}{Provide the distribution of the errors around the mean partner age.
Choose from "gamma" or "normal" distributions.}

\item{link}{When \code{"gamma"} is chosen as the distribution, the link can be either \code{"identity"} or \code{"log"}.
If the distribution is \code{"normal"}, this has no effect.}

\item{max_age}{The non-inclusive right-hand endpoint of the oldest age group within the model population.
Default is 74, so if the oldest age group begins at 60, the age interval is 60-73. Must be less than or equal to 100.}

\item{age_distribution}{Optional: a vector of length \code{length(seq(min(start_ages), max_age))},
where the \code{i}th entry is the
proportion of the model population with age \code{i}. This vector defines the proportion of the model population with every \emph{age}, not \emph{age group}. If not provided, the 2010-2012 U.K.
life tables are used to estimate the population age distribution (see \link[mixage]{uk_life_table}).}
}
\value{
A list, where MOME is the male age mixing matrix, FOME is the female age mixing matrix, AIC is the AIC of the estimated statistical model, and \code{fits} has information on the fit (can be used for sensitivity analysis).
}
\description{
A function to estimate age mixing matrices for user-supplied data. Choose between 
different distributions around mean partner age (Normal or Gamma),
and either an identity or log link for the Gamma distribution. See
"Revisiting Assumptions about Age Preferences in Mathematical 
Models of Sexually Transmitted Infection" (Easterly, et al., 2018) for details
about the estimation procedures. For a function 
to estimate the best mixing structure for your data, see \link[mixage]{best_age_mixing}
}
\examples{
data("mixage_sample_data")
agemix <- estimate_age_mixing(choice_data = mixage_sample_data, 
          start_ages = seq(12, 72, by = 2),
          max_age = 74)
 

}
