% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_age_mixing.R
\name{fill_omegas}
\alias{fill_omegas}
\title{Fill mixing matrices based on estimates.
Mostly private function - only used in sensitivity analysis.}
\usage{
fill_omegas(distribution, fits, start_ages, max_age)
}
\description{
Fill mixing matrices based on estimates.
Mostly private function - only used in sensitivity analysis.
}
