% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixage_sample_data.R
\docType{data}
\name{mixage_sample_data}
\alias{mixage_sample_data}
\title{Sample data.}
\format{An object of class \code{data.frame} with 1000 rows and 3 columns.}
\usage{
mixage_sample_data
}
\description{
A dataframe, with columns \code{chsage}, \code{ptage}, and \code{sex}. Created using the file \code{R/create_sample_data.R}.
}
\keyword{datasets}
