% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_life_table.R
\docType{data}
\name{uk_life_table}
\alias{uk_life_table}
\title{Life table data from UK, 2010 - 2012}
\format{An object of class \code{data.frame} with 101 rows and 11 columns.}
\source{
https://www.ons.gov.uk/peoplepopulationandcommunity/birthsdeathsandmarriages/lifeexpectancies/datasets/nationallifetablesunitedkingdomreferencetables
}
\usage{
uk_life_table
}
\description{
Life table data from UK, 2010 - 2012
}
\keyword{datasets}
