% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_single_year_age_distribution.R
\name{calculate_single_year_age_distribution}
\alias{calculate_single_year_age_distribution}
\title{Calculate age distribution.}
\usage{
calculate_single_year_age_distribution(min_age, max_model_age,
  death_prob = NULL)
}
\arguments{
\item{min_age}{minimum model age}

\item{max_age}{maximum model age}

\item{death_prop}{Optional: a length 99 vector of annual probabilities of death for 1-year-olds to 99-year-olds}
}
\value{
\code{age_prop}: the proportion of the model population with each age
}
\description{
If you have annual probabilities of death (from life tables),
you can use this to calculate an age distribution for \link[mixage]{estimate_age_mixing}
}
