% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_age_mixing.R
\name{estimate_age_mixing}
\alias{estimate_age_mixing}
\title{calculate age mixing structures}
\usage{
estimate_age_mixing(choice_data, max_age, start_ages,
  distribution = c("gamma", "laplace", "normal"), variance_model = c("sqrt",
  "log", "linear", "const"), death_prob = NULL)
}
\arguments{
\item{choice_data}{a dataframe with chooser age, partner age, sex, and optional survey weights. 
These columns should be named \code{chsage}, \code{ptage}, \code{sex}, and \code{weights}, respectively.
Sex must be coded as "Male" and "Female"}

\item{max_age}{The non-inclusive right-hand endpoint of the oldest age group within the model population.
Default is 74, so if the oldest age group begins at 60, the age interval is 60-73. Must be less than or equal to 100.}

\item{start_ages}{vector of the youngest ages included in each age group. 
If \code{start_ages} is \code{c(12, 20, 30)}, the age groups are 12-19, 20-29, and 30 to \code{max_age - 1}.}

\item{distribution}{Provide the distribution of the errors around the mean partner age. Choose from Gamma, Laplace, and Normal distributions.}

\item{variance_model}{describes the transformation of the residuals for estimating the heteroscedastic variance. 
Choose \code{linear}, \code{sqrt}, \code{log}, or \code{const}.}

\item{death_prob}{Length 100 vector. \code{death_prob[1]} is probability of dying in the next year for 1-year-old,
\code{death_prob[12]} is for 12-year-old, and so on. If not supplied, default is taken from  2011 US life tables}
}
\description{
DESCRIBE TRANSFORMATIONS HERE
}
